+++ packages/logging-models/src/connectors/multiLoggingConnector.ts
@@ -1,8 +1,8 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { BaseError, Guards, NotImplementedError } from "@gtsc/core";
-import type { EntityCondition, SortDirection } from "@gtsc/entity";
-import { nameof } from "@gtsc/nameof";
+import { BaseError, Guards, NotImplementedError } from "@twin.org/core";
+import type { EntityCondition, SortDirection } from "@twin.org/entity";
+import { nameof } from "@twin.org/nameof";
 import { LoggingConnectorFactory } from "../factories/loggingConnectorFactory";
 import type { ILogEntry } from "../models/ILogEntry";
 import type { ILoggingConnector } from "../models/ILoggingConnector";