+++ packages/logging-models/src/helpers/logEntryHelper.ts
@@ -1,7 +1,7 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
 
-import { I18n, Is, StringHelper } from "@gtsc/core";
+import { I18n, Is, StringHelper } from "@twin.org/core";
 import type { ILogEntry } from "../models/ILogEntry";
 
 /**