+++ packages/logging-service/tests/loggingService.spec.ts
@@ -1,6 +1,6 @@
 // Copyright 2024 IOTA Stiftung.
 // SPDX-License-Identifier: Apache-2.0.
-import { LoggingConnectorFactory, type ILoggingConnector } from "@gtsc/logging-models";
+import { LoggingConnectorFactory, type ILoggingConnector } from "@twin.org/logging-models";
 import { LoggingService } from "../src/loggingService";
 
 describe("LoggingService", () => {